fix: Fixed a desync issue when I started updating the PlayerState Position
