+++ packages/frontend/src/game/lobby/CreateLobby.tsx
@@ -16,6 +16,7 @@ import { setupCombat } from "./handleCombat";
 import { handleMultiplayer } from "./handleMultiplayer";
 import { AnimationHandler } from "./handleAnimations";
 import { PlayerState } from "./PlayerState";
+import { createLobbyMap } from "./createLobbyMap";
 
 // https://kenney.nl/assets/animated-characters-2
 // mixamo
@@ -86,23 +87,7 @@ const CreateLobby = (): JSX.Element => {
                 }
 
                 // Setup scene lighting and ground
-                const light = new HemisphericLight("light", new Vector3(0, 1, 0), scene);
-                light.intensity = 0.5;
-                const groundOptions = {
-                    width: 1000,
-                    height: 1000,
-                    subdivisions: 10,
-                    subdivisionsX: 10,
-                }   
-                const ground = CreateGround("ground", groundOptions, scene);
-                ground.position.y = 0;
-                const groundMaterial = new StandardMaterial("groundMaterial", scene);
-                groundMaterial.diffuseColor = new Color3(0.5, 0.5, 0.5);
-                ground.material = groundMaterial;
-
-                const wall = createWall(10, 10, 10, 10, 10, scene);
-                wall.position.y = 0;
-                wall.showBoundingBox = true;
+                createLobbyMap(scene);
 
                 // Start rendering
                 scene.executeWhenReady(() => {
